Extract test for confirmation
