+++ spec/features/order_spec.rb
@@ -4,16 +4,6 @@ RSpec.feature 'Order', type: :feature do
   given!(:lunchbox) { create(:lunchbox) }
   given(:order) { create(:order) }
 
-  feature '予約の確認' do
-    scenario '予約者は自分の予約を確認できる' do
-      order_item = create(:order_item, order: order, lunchbox: lunchbox)
-
-      visit order_order_items_path(order)
-      expect(page).to have_text(order_item.customer_name)
-      expect(page).to have_text('2017/02/01(水)の予約一覧')
-    end
-  end
-
   feature '予約の新規作成' do
     scenario 'Order が締め切られている場合、予約者は新しく予約できない' do
       order = create(:order, :closed)
